More on graphql

+++ Udemy/the_modern_graphql_bootcamp/graphql-prisma/src/resolvers/Mutation.js
@@ -1,30 +1,75 @@
-import uuidv4 from 'uuid/v4'
-import prisma from '../prisma'
+import bcrypt from 'bcryptjs'
+
+import getUserId from '../utils/getUserId'
+import generateToken from '../utils/generateToken'
+import hashPassword from '../utils/hashPassword'
 
 const Mutation = {
     async createUser(parent, args, { prisma }, info) {
-        if (args.data.password.length < 8) {
-            throw new Error('Password must be 8 characters or longer.')
+        const password = await hashPassword(args.data.password)
+
+        const user = prisma.mutation.createUser({
+            data: {
+                ...args.data,
+                password
+            }
+        })
+
+        return {
+            user,
+            token: generateToken(user.id)
+        }
+    },
+    async login(parent, args, { prisma }, info) {
+        const { email, password } = args.data
+
+        const user = await prisma.query.user({
+            where: {
+                email
+            }
+        })
+
+        if (!user) {
+            throw new Error('Unable to login')
         }
 
-        return prisma.mutation.createUser({ data: args.data }, info)
+        const isMatch = await bcrypt.compare(password, user.password)
+
+        if (!isMatch) {
+            throw new Error('Unable to login')
+        }
+
+        return {
+            user,
+            token: generateToken(user.id)
+        }
     },
-    async deleteUser(parent, args, { prisma }, info) {
+    async deleteUser(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+
         return prisma.mutation.deleteUser({ 
             where: {
-                id: args.id 
+                id: userId
             }
         }, info)
     },
-    async updateUser(parent, args, { db }, info) {
+    async updateUser(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+
+        if (typeof args.data.password === 'string') {
+            args.data.password = await hashPassword(args.data.password)
+        }
+
         return prisma.mutation.updateUser({
             where: {
-                id: args.id
+                id: userId
             },
             data: args.data
         }, info)
     },
-    async createPost(parent, args, { prisma }, info) {
+    async createPost(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+
         return prisma.mutation.createPost({
             data: {
                 title: args.data.title,
@@ -32,20 +77,55 @@ const Mutation = {
                 published: args.data.published,
                 author: {
                     connect: {
-                        id: args.data.author 
+                        id: userId
                     }
                 }
             }
         }, info)
     },
-    async deletePost(parent, args, { prisma }, info) {
+    async deletePost(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+        const postExists = await prisma.exists.Post({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+
+        if (!postExists) {
+            throw new Error('Unable to delete post')
+        }
+
         return prisma.mutation.deletePost({
             where: {
                 id: args.id
             }
         }, info)
     },
-    async updatePost(parent, args, { prisma }, info) {
+    async updatePost(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+        const postExists = await prisma.exists.Post({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+        const idPublished = await prisma.exists.Post({ id: args.id, published: true })
+
+        if (!postExists) {
+            throw new Error('Unable to update post')
+        }
+
+        if (idPublished && args.data.published === false) {
+            await prisma.mutation.deleteManyComments({
+                where: {
+                    post: {
+                        id: args.id
+                    }
+                }
+            })
+        }
+
         return prisma.mutation.updatePost({
             where: {
                 id: args.id
@@ -53,13 +133,23 @@ const Mutation = {
             data: args.data
         }, info)
     },
-    async createComment(parent, args, { prisma }, info) {
+    async createComment(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+        const postExists = await prisma.exists.Post({
+            id: args.data.post,
+            published: true
+        })
+
+        if (!postExists) {
+            throw new Error('Unable to create comment. Post not found')
+        }
+
         return prisma.mutation.createComment({
             data: {
                 text: args.data.text,
                 author: {
                     connect: {
-                        id: args.data.author
+                        id: userId
                     }
                 },
                 post: {
@@ -70,14 +160,38 @@ const Mutation = {
             }
         }, info)
     },
-    async deleteComment(parent, args, { prisma }, info) {
+    async deleteComment(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+        const commentExists = await prisma.exists.Comment({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+
+        if (!userId) {
+            throw new Error('Unable to delete comment')
+        }
+
         return prisma.mutation.deleteComment({
             where: {
                 id: args.id
             }
         }, info)
     },
-    async updateComment(parent, args, { prisma }, info) {
+    async updateComment(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+        const commentExists = await prisma.exists.Comment({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+
+        if (!commentExists) {
+            throw new Error('Unable to update comment')
+        }
+
         return prisma.mutation.updateComment({
             where: {
                 id: args.id
